requirements.txt
coverage==4.5.3
dj-database-url==0.5.0
Django==2.1.7
django-activity-stream==0.7.0
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-extra-views==0.12.0
django-filter==2.1.0
django-imagekit==4.0.2
django-markdown-deux==1.0.5
django-pagedown==1.0.6
django-taggit==1.1.0
django-widget-tweaks==1.4.3
Markdown==3.0.1
markdown2==2.3.7
np==1.0.2
numpy==1.16.2
pilkit==2.0
Pillow==5.4.1
pinax-blog==7.0.4
pinax-images==3.0.1
pinax-likes==3.0.1
pinax-messages==2.0.2
pinax-notifications==5.0.3
pinax-testimonials==2.0.4
Pygments==2.3.1
python-decouple==3.1
pytz==2018.9
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
Pygments==2.3.1
python-decouple==3.1
pytz==2018.9
selenium==3.141.0
six==1.12.0
urllib3==1.24.1